pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrdlp-lab"
version = "0.1.0"
description = "OCR+DLP dataset collection and annotation tool chain"
authors = [{name = "OCR Lab", email = "lab@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.1",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.1.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "openai>=1.3.7",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "imagehash>=4.3.1",
    "click>=8.1.7",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.6",
    "black>=23.11.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[project.scripts]
ocrdlp = "cli:cli"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.black]
target-version = ['py311']
line-length = 100
skip-string-normalization = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function" 